fix(p5): phpstan L5 — zero errors
Climbs the PHPStan level from 4 to 5, fixing every argument.type
fallout at the root cause.

Genuine bugs found and fixed:
- functions_metadata.php: checkdate()/round() called with preg_match()
  string captures and SVG viewBox tokens under strict_types=1 — both
  would fatal (TypeError) the instant this code path actually ran; cast
  to int/float at the call site.
- functions.inc.php/admin/include/functions.php: two
  preg_replace_callback() closures declared `: int` return types (one
  from ord(), one from an array lookup) but the callback contract
  requires string — cast both to string.
- base32.class.php: encode()/decode() called base_convert()/chr()/ord()
  with mismatched string/int args — genuinely fatal under
  strict_types=1 (verified: reverting the fix reproduces the crash).
  Confirmed the encode/decode round-trip still works for the 8-char/
  no-padding case actually used by totp.class.php (20-byte secrets,
  a clean multiple of 5). Also found, but explicitly left alone (out of
  this pass's scope — not a type error, a separate pre-existing logic
  bug never previously reachable since encode() couldn't even run):
  decode() corrupts its output with trailing NUL bytes for any input
  whose length isn't a multiple of 5 bytes, because missing/
  out-of-range flippedMap lookups silently zero-pad instead of being
  excluded. Harmless today only because the one real caller always
  uses 20-byte secrets.
- tools/test_piwigo.php: its own local log_user(array $option, $cookies)
  helper collided with the real global log_user($user_id, $remember_me)
  in functions_user.inc.php — same name, both in the global namespace.
  PHPStan resolved a real call site (register.php) against the wrong
  one. Renamed the CLI script's helper to test_log_user().
- functions_cookie.inc.php/picture.php: setcookie(..., false, ...) to
  clear a cookie — false isn't a valid $value, silently worked before
  strict_types instrumentation caught it; changed to ''.
- functions_mysqli.inc.php: pwg_db_connect()'s host:port split passed
  the port half of explode(':', $host) straight to mysqli's constructor
  as a string; cast to int.
- comments.php: create_navigation_bar()'s 5th arg was '' where every
  other call site passes true or omits it for the boolean default —
  removed the stray argument.
- tools/test_piwigo.php: PHPWG_INSTALLED === true can never
  distinguish installed/not-installed (already fixed in L4, this level
  cleaned up 4 CURLOPT_POST => 1 to => true for curl_setopt_array's
  bool-typed option shape).

Docblock widenings backed by real, verified call sites: format_date()
(DateTime|false passthrough from format_fromto()), userprefs_update_param
(bool/int/array real callers), calculate_auto_login_key/log_user
(numeric-string cookie fields), set_tags/add_tags/update_category/
find_tags/get_image_ids_for_tags (explode()'d numeric-string tag/category
ids, only ever used in implode()/SQL contexts), get_db_plugins (null
same as '' via empty()), page_not_found (null same as '' via
concatenation), ScriptLoader::add() and SizingParams::compute() (null
sentinels already handled by empty()-based logic further down),
delete_element_derivatives (IMG_CUSTOM's string value is a legitimate
type, not just the numeric IMG_* constants), template.class.php's
derivative() (fixed to wrap array input in `new SrcImage()` like its
sibling derivative_url()/DerivativeImage::url() already do — an
unreachable-until-now bug, no template currently calls it).

Tooling: added ecs.php to phpstan.neon's excludePaths — it references
PhpCsFixer\Fixer\...\*Fixer::class constants that only exist nested
inside symplify/easy-coding-standard's bundled vendor tree, not composer.json's
own dependency graph; ECS resolves and runs it fine (composer lint:php
stays green throughout), but our root autoloader can't see those classes,
so PHPStan can't type the ::class references as class-string<FixerInterface>.
Same rationale as the existing tools/phpstan-bootstrap.php exclusion.

Full suite green: PHPStan 0 errors, ECS clean, Unit 10, Integration 3,
Contract 93, Browser 44, Visual 32 (two transient VR flakes during this
pass — investigated via extracted pixel diffs, confirmed as pre-existing
sub-pixel AA jitter and an image-load timing race, not tied to any file
touched this level; a subsequent clean run showed 32/32 passing).

+++ admin/include/functions_metadata.php
@@ -28,9 +28,9 @@ function get_sync_iptc_data($file): array
     foreach ($iptc as $pwg_key => $value) {
         if (in_array($pwg_key, ['date_creation', 'date_available'])) {
             if (preg_match('/(\d{4})(\d{2})(\d{2})/', (string) $value, $matches)) {
-                $year = $matches[1];
-                $month = $matches[2];
-                $day = $matches[3];
+                $year = (int) $matches[1];
+                $month = (int) $matches[2];
+                $day = (int) $matches[3];
 
                 if (! checkdate($month, $day, $year)) {
                     // we suppose the year is correct
@@ -178,13 +178,13 @@ function get_sync_metadata($infos)
                 if (isset($width) and $width != '') {
                     $infos['width'] = (int) $width;
                 } else {
-                    $infos['width'] = round(explode(' ', $vb)[2]);
+                    $infos['width'] = round((float) explode(' ', $vb)[2]);
                 }
 
                 if (isset($height) and $height != '') {
                     $infos['height'] = (int) $height;
                 } else {
-                    $infos['height'] = round(explode(' ', $vb)[3]);
+                    $infos['height'] = round((float) explode(' ', $vb)[3]);
                 }
             }
             if ($image_size = @getimagesize($file)) {
